+++ src/main/java/com/mink/freshexpress/order/model/StockReservation.java
@@ -9,7 +9,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -22,7 +22,7 @@ public class StockReservation extends BaseEntity {
     private Long id;
 
     @Column(nullable = false)
-    private BigInteger quantity;
+    private BigDecimal quantity;
 
     @Enumerated(EnumType.STRING)
     private ReservationStatus status;
